Refactor EffectServo and EffectStrobe to use optimized math - EffectServo: Replaced floating-point arithmetic with 8.8 fixed-point (uint16_t). - EffectStrobe: Replaced expensive modulo operator with conditional subtraction. - Updated Effect.h to reflect type changes.

+++ src/effects/Effect.h
@@ -96,9 +96,9 @@ public:
 private:
     uint8_t _endpoint_a;
     uint8_t _endpoint_b;
-    float _speed;
-    float _current_angle;
-    float _target_angle;
+    uint16_t _speed; // Fixed point 8.8 (units per ms)
+    uint16_t _current_angle; // Fixed point 8.8
+    uint16_t _target_angle; // Fixed point 8.8
     bool _is_at_a = true;
 };
